+++ packages/project-builder/src/ProjectBuilderStore.ts
@@ -627,6 +627,10 @@ export function createProjectBuilderSlice<PC extends BaseProjectConfig>(
             const url = ds.url;
             setProjectFileProgress(fileName, {status: 'download'});
             const downloadedFile = await downloadFile(url, {
+              ...(ds.type === DataSourceTypes.Enum.url && {
+                method: ds.httpMethod,
+                headers: ds.headers,
+              }),
               onProgress: (progress: ProgressInfo) => {
                 setProjectFileProgress(fileName, {
                   status: 'download',