+++ packages/ui/src/components/theme-switch.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import * as React from 'react';
 import {useTheme} from '../theme/theme-provider';
 import * as SwitchPrimitives from '@radix-ui/react-switch';