+++ packages/jest-config/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.10.0 (2025-04-08)
+
+**Note:** Version bump only for package @sqlrooms/jest-config
+
+## <small>0.9.3 (2025-04-03)</small>
+
+## <small>0.9.2 (2025-04-01)</small>
+
+## <small>0.9.1 (2025-04-01)</small>
+
+# [0.9.0](https://github.com/sqlrooms/sqlrooms/compare/v0.8.1...v0.9.0) (2025-03-31)
+
+**Note:** Version bump only for package @sqlrooms/jest-config
+
 ## [0.8.1](https://github.com/sqlrooms/sqlrooms/compare/v0.8.0...v0.8.1) (2025-03-10)
 
 **Note:** Version bump only for package @sqlrooms/jest-config