+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.10.0 (2025-04-08)
+
+**Note:** Version bump only for package @sqlrooms/ui
+
+## <small>0.9.3 (2025-04-03)</small>
+
+- fix: Properly map all @sqlrooms/ui subpath exports (#38) ([2edd7d8](https://github.com/sqlrooms/sqlrooms/commit/2edd7d8)), closes [#38](https://github.com/sqlrooms/sqlrooms/issues/38)
+
+## <small>0.9.2 (2025-04-01)</small>
+
+## <small>0.9.1 (2025-04-01)</small>
+
+# [0.9.0](https://github.com/sqlrooms/sqlrooms/compare/v0.8.1...v0.9.0) (2025-03-31)
+
+**Note:** Version bump only for package @sqlrooms/ui
+
 ## [0.8.1](https://github.com/sqlrooms/sqlrooms/compare/v0.8.0...v0.8.1) (2025-03-10)
 
 **Note:** Version bump only for package @sqlrooms/ui